굼바 시야 범위 안에 들어왔을때만 플레이어 찾도록 함
- gameobj에 시야 안에 들어왔는지 확인하는 함수 추가

- float에 double넣는 코드들 수정

+++ Engine_SOURCE/LifeUIScript.cpp
@@ -145,9 +145,9 @@ void LifeUIScript::BackUI()
 
 	Vector3 dir = mOriginPos - mTargetPos;
 
-	Vector3 fabsDir = Vector3(dir.x, dir.y, 0);
-	double value = sqrt(pow(fabsDir.x, 2) + pow(fabsDir.y, 2));
-	Vector3 dirValue = Vector3(fabsDir.x / value, fabsDir.y / value, 0);
+	Vector3 fabsDir = Vector3(dir.x, dir.y, 0.f);
+	float value = sqrt(pow(fabsDir.x, 2) + pow(fabsDir.y, 2));
+	Vector3 dirValue = Vector3(fabsDir.x / value, fabsDir.y / value, 0.f);
 
 	Vector3 pos = mTr->GetPosition();
